app refactoring
